Report missing digitization_work_order dependency more clearly
There was already some code in there to show a nice message, but it
was failing because the `require "write_xlsx"` was getting loaded
first and throwing an error.

We don't need the require anyway, since Bundler will do that, so I've
removed it and smartened up the message a bit.